requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-classic
langchain-chroma
chromadb
sentence-transformers
duckduckgo-search
ddgs
scikit-learn
numpy
xgboost
pandas
pydantic
python-dotenv
